[storage] No display secret
