+++ src/moonlink/src/storage.rs
@@ -15,7 +15,8 @@ pub(crate) use cache::object_storage::cache_handle::NonEvictableHandle;
 pub use cache::object_storage::object_storage_cache::ObjectStorageCache;
 pub use filesystem::accessor::filesystem_accessor::FileSystemAccessor;
 pub use filesystem::filesystem_config::FileSystemConfig;
-pub use iceberg::iceberg_table_manager::{IcebergTableConfig, IcebergTableManager};
+pub use iceberg::iceberg_table_config::IcebergTableConfig;
+pub use iceberg::iceberg_table_manager::IcebergTableManager;
 pub use iceberg::table_event_manager::{EventSyncReceiver, TableEventManager};
 pub use iceberg::table_manager::TableManager;
 pub use mooncake_table::table_config::TableConfig as MoonlinkTableConfig;
